refactor: Move IntoOption trait to the only file where it's used
